Changed build system to use maven-nar-plugin to compile the binaries and make it reusable for Maven builds - Added a test to the Maven build
