--- .cvsignore
@@ -1,7 +1,0 @@
-Arch
-jshortcut.dll
-jshortcut.jar
-mkl
-obj
-tmp
-tmp?
