--- README.build
@@ -1,39 +1,0 @@
-JShortcut					September 12, 2003
-
-Building a new version
-----------------------
-
-    Check these files to make sure they contain the latest dates, version
-    numbers, and other information:
-
-	COPYRIGHT		Copyright date, file date
-	DONE			Version and date
-	HISTORY			Release notes for this version, date
-	README.build		Date
-	README.html		Version and date
-	VERSION			Version
-
-    The Makefiles are set up for GNU make.  You can build everything
-    except for the DLL on any machine with Java and GNU make.  You can
-    build the DLL separately on a Windows machine by running the
-    compile.bat script in the src/jni directory.  Do this just after
-    doing the "make all" in the sequence below.
-
-    In the top level directory, execute these commands:
-
-	make allclean		Clean up before building
-        make all		Builds everything (jar, apidoc)
-	make rel		Builds a release kit named jshortcut-<version>
-	make reljar		Builds a self-extracting jar file
-					from the release kit
-
-    If you are maintaining archives and CVS:
-
-	make arch		Copy the jar & README files to archive dir
-	make cvstag		Run "cvs tag" for this version
-
-    To clean up, execute this command:
-
-        make allclean		Deletes all generated files
-
-    For more detailed build and clean options, see the Makefile.
