--- misc/manifest.mf
@@ -1,1 +1,0 @@
-Main-class: net.jimmc.jshortcut.JShellLink
