+++ pom.xml
@@ -1,0 +1,96 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<groupId>net.jimmc</groupId>
+	<artifactId>jshortcut</artifactId>
+	<version>1.0.1</version>
+	<properties>
+		<sonar.junit.reportsPath>test-output/junitreports</sonar.junit.reportsPath>
+		<timestamp>${maven.build.timestamp}</timestamp>
+		<maven.build.timestamp.format>yyyyMMdd</maven.build.timestamp.format>
+	</properties>
+	<packaging>nar</packaging>
+	<distributionManagement>
+		<repository>
+			<id>soprasdlc</id>
+			<name>Sopra SDLC repository internal</name>
+			<url>http://slsdlc:8081/nexus/content/repositories/releases/</url>
+		</repository>
+	</distributionManagement>
+	<dependencies>
+        <dependency>
+            <groupId>org.scijava</groupId>
+            <artifactId>native-lib-loader</artifactId>
+            <version>2.0.2</version>
+        </dependency>
+        <dependency>
+        	<groupId>junit</groupId>
+        	<artifactId>junit</artifactId>
+        	<version>4.12</version>
+        </dependency>
+  	</dependencies>
+	<build>
+		<plugins>
+		<plugin>
+		    <groupId>org.codehaus.mojo</groupId>
+		    <artifactId>build-helper-maven-plugin</artifactId>
+		    <executions>
+		        <execution>
+		            <id>add-source</id>
+		            <phase>generate-sources</phase>
+		            <goals>
+		                <goal>add-source</goal>
+		            </goals>
+		            <configuration>
+		                <sources>
+		                    <source>${project.build.directory}/nar/nar-generated/</source>
+		                </sources>
+		            </configuration>
+		        </execution>
+		    </executions>
+		 </plugin>
+			<plugin>
+				<groupId>com.github.maven-nar</groupId>
+				<artifactId>nar-maven-plugin</artifactId>
+				<version>3.3.0</version>
+				<extensions>true</extensions>
+				<configuration>
+					<runtime>static</runtime>
+					<cpp>
+						<multiThreaded>true</multiThreaded>
+					</cpp>
+					<linker>
+						<name>msvc</name>
+						<sysLibs>
+							<sysLib>
+								<name>advapi32</name>
+							</sysLib>
+							<sysLib>
+								<name>shell32</name>
+							</sysLib>
+							<sysLib>
+								<name>ole32</name>
+							</sysLib>
+							<sysLib>
+								<name>uuid</name>
+							</sysLib>
+							<sysLib>
+								<name>kernel32</name>
+							</sysLib>
+							<sysLib>
+								<name>libucrt</name>
+							</sysLib>
+						</sysLibs>
+					</linker>
+					<libraries>
+						<library>
+							<narSystemPackage>net.jimmc.jshortcut</narSystemPackage>
+							<type>jni</type>
+						</library>
+					</libraries>
+				</configuration>
+			</plugin>
+		</plugins>
+	
+	</build>
+</project>
