--- src/jni/jshortcut.def
@@ -1,16 +1,0 @@
-; JShortcut DLL
-;
-
-LIBRARY   JSHORTCUT
-CODE      PRELOAD MOVEABLE DISCARDABLE
-DATA      PRELOAD MOVEABLE SINGLE
-HEAPSIZE  8192
-
-; Exports for JShortcut
-EXPORTS
-
-	Java_net_jimmc_jshortcut_JShellLink_nGetDirectory  @10
-	Java_net_jimmc_jshortcut_JShellLink_nLoad          @11
-	Java_net_jimmc_jshortcut_JShellLink_nSave          @12
-
-;
