--- src/net/jimmc/jshortcut/.cvsignore
@@ -1,3 +1,0 @@
-mkl
-tmp
-tmp?
