--- src/net/jimmc/jshortcut/Makefile
@@ -1,3 +1,0 @@
-TOPDIR = ../../../..
-MAKEINFODIR = $(TOPDIR)/makeinfo
-include $(MAKEINFODIR)/javapkg.mak
