--- src/net/jimmc/selfunzip/Makefile
@@ -1,5 +1,0 @@
-TOPDIR = ../../../..
-MAKEINFODIR = $(TOPDIR)/makeinfo
-NOJAR = true
-
-include $(MAKEINFODIR)/javapkg.mak
