--- src/net/jimmc/selfunzip/manifest.mf
@@ -1,1 +1,0 @@
-Main-class: net.jimmc.selfunzip.ZipSelfExtractor
